First Review: Added Heteroscedasticity in Regression as assigned in issue #166
Added Heteroscedasticity in Regression

The tutorial begins with a short summary about heteroscedasticity and its significance. Next, we use the residual plot code to detect heteroscedasticity with the naked eye and provide a summary of how the plot can be characterized by heteroscedasticity. Lastly, we provide a remedy of fixing heteroscedasticity by transforming the dependent variable with the help of a Box-Cox transformation and visualizing the residual plot again.